Clarify distinction between buffered and unbuffered comments.
The Comments page already described the `//-` tag as an "unbuffered"
comment, so I thought it would help to state the reverse -- that `//`
lines and blocks are "buffered" comments. Inspired by an issue on the
Pug core repo, "Top-level comments in extending templates", where the
issue reporter was apparently confused by the difference between the
two.

https://github.com/pugjs/pug/issues/2632

This will need to be updated if anyone implements the feature suggested
in that thread, where even buffered comments at the top level of an
extending template are simply ignored.