Expand notes about plain text, inline-level elements, and whitespace.
I rearranged the Plain Text page partly because I disagree with the
statement that piped text is the simplest way of adding plain text.
Whitespace for plain text is pretty tricky for beginners, and I think
piped text in particular seems to come up a lot as a source of
confusion.

So for the Plain Text page, I:

* rearranged the order.
* pointed out the the "weird" whitespace behavior is actually a benefit
for template writers.
* noted that starting a line with literal HTML is also plain text.
* expanded the section about dot-blocks of plain text.
* and wrote a whole section specifically about different ways to control
whitespace in the rendered HTML.

As for the other pages, I mostly just added cross-references where plain
text and whitespace was related. This seems like an important enough
point that we should make sure readers can easily find their way back
here.