Magical mistakes and how to avoid them (#11)
* Remove "extends" page (#9)

All of this is covered in more detail on the "Inheritance" page.

* Note common mistakes when working with template inheritance.

Inspired by my issue on the core Pug repo, "Error when there's
unbuffered code at the top of an extending template".

https://github.com/pugjs/pug/issues/2721

In that issue, I guess I'd forgotten that unbuffered code can also
contain Pug markup, because I usually do well enough designing my
template logic around the built-in `if`, `unless`, and `each` keywords.

* Clarify distinction between buffered and unbuffered comments.

The Comments page already described the `//-` tag as an "unbuffered"
comment, so I thought it would help to state the reverse -- that `//`
lines and blocks are "buffered" comments. Inspired by an issue on the
Pug core repo, "Top-level comments in extending templates", where the
issue reporter was apparently confused by the difference between the
two.

https://github.com/pugjs/pug/issues/2632

This will need to be updated if anyone implements the feature suggested
in that thread, where even buffered comments at the top level of an
extending template are simply ignored.

* Expand notes about plain text, inline-level elements, and whitespace.

I rearranged the Plain Text page partly because I disagree with the
statement that piped text is the simplest way of adding plain text.
Whitespace for plain text is pretty tricky for beginners, and I think
piped text in particular seems to come up a lot as a source of
confusion.

So for the Plain Text page, I:

* rearranged the order.
* pointed out the the "weird" whitespace behavior is actually a benefit
for template writers.
* noted that starting a line with literal HTML is also plain text.
* expanded the section about dot-blocks of plain text.
* and wrote a whole section specifically about different ways to control
whitespace in the rendered HTML.

As for the other pages, I mostly just added cross-references where plain
text and whitespace was related. This seems like an important enough
point that we should make sure readers can easily find their way back
here.

* Keep trailing whitespace in Pug and Markdown files.

Trailing whitespace is significant in lines and blocks of plain text and
inline-level tags in Pug. Markdown files that have Pug examples should
therefore also preserve trailing whitespace.

* For tag interpolation add an example containing an attribute.

When I was first getting my head around Pug, I wasn't quite sure where
at first if tag attributes would work or where they should go, in an
interpolated tag.

* Sort Whitespace Control solutions into recommended and not recommended.

`= "\n"` is actually an example of buffered code, not string
interpolation.

* Remove examples of buffered code used as a tag separator.

It was maybe kind of a weird hack anyway.

* Pretty mode is deprecated.

My note about the whitespace behavior for inline-leve tags is actually
also true about block-level tags, except when running in "pretty" mode,
which has been a common source of bugs in users' templates.

* Really flesh out the rules for when whitespace is removed or preserved.

Also, if we're actively discouraging people from using pretty mode,
maybe I guess we shouldn't bother with allusions to removing whitespace
"by default." They're better off not enabling pretty mode to put
whitespace back in since it's inconsistent.

* grammar

* one of the trickiest parts