Fix danger section
Here is a quick fix for the missing danger section style. It should be handled in the stylesheet by editing .alert.alert-danger style but it seems this stylesheet is not open-source. So this worked in my live tests on the page. If the markdown parser does not strip it, it should work.